+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 ![DIIVE](images/logo_diive1_256px.png)
 
+## v0.79.1 | 26 Aug 2024
+
+### Additions
+
+- Added new function to apply quality flags to certain time periods only (`diive.pkgs.qaqc.flags.restrict_application`)
+- Added to option to restrict the application of the angle-of-attack flag to certain time periods (
+  `diive.pkgs.fluxprocessingchain.level2_qualityflags.FluxQualityFlagsEddyPro.angle_of_attack_test`)
+
+### Changes
+
+- Test options in `FluxProcessingChain` are now always passed as dict. This has the advantage that in addition to run
+  the test by setting the dict key `apply` to `True`, various other test settings can be passed, for example the new
+  parameter `application dates` for the angle-of-attack flag. (
+  `diive.pkgs.fluxprocessingchain.fluxprocessingchain.FluxProcessingChain`)
+
+### Tests
+
+- Added unittest for Flux Processing Chain up to Level-2 (
+  `tests.test_fluxprocessingchain.TestFluxProcessingChain.test_fluxprocessingchain_level2`)
+- 36/36 unittests ran successfully
+
 ## v0.79.0 | 22 Aug 2024
 
 This version introduces a histogram plot that has the option to display z-score as vertical lines superimposed on the
@@ -23,7 +44,8 @@ time series. The bin with most counts is highlighted orange.*
 
 ### Additions
 
-- Added histogram plots to `FlagBase`, histograms are now shown for all outlier methods (`diive.core.base.flagbase.FlagBase.defaultplot`)
+- Added histogram plots to `FlagBase`, histograms are now shown for all outlier methods (
+  `diive.core.base.flagbase.FlagBase.defaultplot`)
 - Added daytime/nighttime histogram plots to (`diive.pkgs.outlierdetection.hampel.HampelDaytimeNighttime`)
 - Added daytime/nighttime histogram plots to (`diive.pkgs.outlierdetection.zscore.zScoreDaytimeNighttime`)
 - Added daytime/nighttime histogram plots to (`diive.pkgs.outlierdetection.lof.LocalOutlierFactorDaytimeNighttime`)